Add support for more types of leaderboards
With the 1.16 update of Minecraft, netherite is the new most valuable item,
and the existing leaderboard systems don't allow players to compete in a
meaningful way for netherite acquisition.

This patch tracks 3 new events which are attempt to address this gap in
different ways.

Three new config options are added:
* record-furnace-smelting - count smelted items removed from furnaces
* record-crafting - count items crafted on workbenches and crafting grid
* record-smithing - count items crafted on smithing tables

The furnace smelting option allows counting the process of smelting
ancient debris to netherite scraps. Due to limitations in bukkit, this
event doesn't fire in every case you would want, and thus a leaderboard
based on it can undercount, but it's still valuable.

The crafting option allows reliable counting of creation of netherite
ingots from scraps, although the code is somewhat complex. The crafting
option also allows creation of leaderboards for any rare/valuable recipe,
and this patch implements a few others.

The smithing option allows counting of netherite items made from diamond
items plus netherite ingots, to allow for specific leaderboards for
individual tools/armor pieces.